Move reserve component into reservation directory
